backend/pyproject.toml
[project]
name = "opencau-agent-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "httpx==0.28.1",
    "pillow==11.0.0",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "redis==5.2.1",
    "uvicorn[standard]==0.34.0",
    "websockets==14.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
